requirements.txt
asgiref==3.3.4
autopep8==1.5.6
Django==3.2
django-crispy-forms==1.11.2
pycodestyle==2.7.0
python-slugify==4.0.1
pytz==2021.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
